pyproject.toml
[project]
name = "pixelprobe"
version = "0.1.0"
description = "Pixel level image analysis tool with denoising tools and CNN."
authors = [
    {name = "Y. Arin Manav",email = "224327686+arinmanav@users.noreply.github.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "customtkinter (>=5.2.2,<6.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "scikit-image (>=0.25.2,<0.26.0)",
    "scipy (>=1.16.1,<2.0.0)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.ml.dependencies]
torch = "<2.8"
torchvision = "<0.23"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
black = "^25.1.0"
flake8 = "^7.3.0"
mypy = "^1.17.1"
pre-commit = "^4.2.0"
pytest-cov = "^6.2.1"


[tool.poetry.group.extras.dependencies]
h5py = "^3.14.0"
plotly = "^6.2.0"
dash = "^3.2.0"

